Moved electron related directories to new electron directory
