+++ src/electron/main/mainWindow.js
@@ -29,7 +29,7 @@ export const createMainWindow = () => {
     title: 'Hedron',
     // get Hedron icon to appear during dev (only works for win and linux)
     // for better icons, still need to build the app
-    icon: isDevelopment && path.join(__dirname, '../../build/icon.png'),
+    icon: isDevelopment && path.join(__dirname, '../../../build/icon.png'),
     ...dimensions,
   })
