+++ src/electron/main/menu.js
