+++ src/electron/renderer/menuHandler.js
@@ -1,7 +1,7 @@
 import { ipcRenderer } from 'electron'
-import { projectSave, projectSaveAs, projectLoad } from '../store/project/actions'
-import { uiAuxToggleOpen } from '../store/ui/actions'
-import history from '../history'
+import { projectSave, projectSaveAs, projectLoad } from '../../store/project/actions'
+import { uiAuxToggleOpen } from '../../store/ui/actions'
+import history from '../../history'
 
 let dispatch
